+++ frontend/src/pages/postman/PostmanPage.tsx
@@ -1,10 +1,22 @@
 import { Box } from '@chakra-ui/react';
-import { FC, useState } from 'react';
+import { FC, useEffect, useState } from 'react';
 import ResponseList from '../../components/postman/list/ResponseList.tsx';
 import PostmanSidebar from '../../components/postman/sidebar/PostmanSidebar.tsx';
 import EnvironmentList from '../../components/postman/list/EnvironmentList.tsx';
 import PostmanMainSection from '../../components/postman/section/PostmanMainSection.tsx';
 import EnvironmentSection from '../../components/postman/section/EnvironmentSection.tsx';
+import { environmentList } from '../../lib/postmanData.ts';
+
+export const getActiveEnvironmentData = async (environment: string) => {
+  for (const group of environmentList) {
+    for (const envKey in group) {
+      if (envKey === environment) {
+        return group[envKey];
+      }
+    }
+  }
+  return null;
+};
 
 const PostmanPage: FC<{}> = () => {
   // Layout
@@ -18,6 +30,19 @@ const PostmanPage: FC<{}> = () => {
 
   // Collection
   const [activeResponse, setActiveResponse] = useState<string>('Login');
+
+  useEffect(() => {
+    const fetchEnvironmentData = async () => {
+      if (activeEnvironment) {
+        const environmentData = await getActiveEnvironmentData(activeEnvironment);
+        if (environmentData) {
+          setEnvironmentData(environmentData);
+        }
+      }
+    };
+
+    fetchEnvironmentData();
+  }, [activeEnvironment]);
 
   return (
     <Box bg="#222222" minH="100vh" display="flex" gap="1rem">
@@ -44,7 +69,6 @@ const PostmanPage: FC<{}> = () => {
           activeResponseSection={activeResponseSection}
           setActiveResponseSection={setActiveResponseSection}
           activeResponse={activeResponse}
-          activeEnvironment={activeEnvironment}
         />
       )}
